データベースのディレクトリ構成修正
