+++ api/interfaces/controllers/Tags/tag_controller.go
@@ -8,6 +8,7 @@ import (
 
 	controllers "github.com/kory-jp/react_golang_mux/api/interfaces/controllers/sessions"
 	"github.com/kory-jp/react_golang_mux/api/interfaces/database"
+	tags "github.com/kory-jp/react_golang_mux/api/interfaces/database/tags"
 	usecase "github.com/kory-jp/react_golang_mux/api/usecase/tag"
 
 	"github.com/kory-jp/react_golang_mux/api/domain"
@@ -33,7 +34,7 @@ func (res *Response) SetResp(status int, mess string, tags domain.Tags) (resStr 
 func NewTagController(sqlHandler database.SqlHandler) *TagController {
 	return &TagController{
 		Interactor: usecase.TagInteractor{
-			TagRepository: &database.TagRepository{
+			TagRepository: &tags.TagRepository{
 				SqlHandler: sqlHandler,
 			},
 		},
