+++ api/interfaces/database/sessions/mysql/mysql.go
@@ -1,0 +1,19 @@
+package mysql
+
+var FindByEmailState = `
+	select
+		*
+	from
+		users
+	where
+	 email = ?
+`
+
+var FindByIdState = `
+	select
+		*
+	from
+		users
+	where
+	 id = ?
+`
