+++ api/interfaces/database/tags/mysql/mysql.go
@@ -1,0 +1,10 @@
+package mysql
+
+var GetTagsState = `
+	select
+		id,
+		value,
+		label
+	from
+		tags
+`
