+++ api/interfaces/database/todo_tag_relations/mysql/mysql.go
@@ -1,0 +1,18 @@
+package mysql
+
+var CreateTodoTagRelationsState = `
+	insert into
+		todo_tag_relations(
+			todo_id,
+			tag_id,
+			created_at
+		)
+	value (?, ?, ?)
+`
+
+var DeleteTodoTagRelationsState = `
+		delete from
+			todo_tag_relations
+		where
+			todo_tag_relations.todo_id = ?
+`
