+++ api/interfaces/database/todo_tag_relations/todo_tag_relations_repository.go
@@ -5,32 +5,19 @@ import (
 	"fmt"
 	"log"
 	"time"
+
+	"github.com/kory-jp/react_golang_mux/api/interfaces/database/todo_tag_relations/mysql"
+
+	"github.com/kory-jp/react_golang_mux/api/interfaces/database"
 )
 
 type TodoTagRelationsRepository struct {
-	SqlHandler
+	database.SqlHandler
 }
-
-var CreateTodoTagRelationsState = `
-	insert into
-		todo_tag_relations(
-			todo_id,
-			tag_id,
-			created_at
-		)
-	value (?, ?, ?)
-`
-
-var DeleteTodoTagRelationsState = `
-		delete from
-			todo_tag_relations
-		where
-			todo_tag_relations.todo_id = ?
-`
 
 func (repo *TodoTagRelationsRepository) TransStore(tx *sql.Tx, todoId int64, tagIds []int) (err error) {
 	for _, v := range tagIds {
-		_, err = repo.TransExecute(tx, CreateTodoTagRelationsState, todoId, v, time.Now())
+		_, err = repo.TransExecute(tx, mysql.CreateTodoTagRelationsState, todoId, v, time.Now())
 		if err != nil {
 			fmt.Println(err)
 			log.Println(err)
@@ -41,14 +28,14 @@ func (repo *TodoTagRelationsRepository) TransStore(tx *sql.Tx, todoId int64, tag
 }
 
 func (repo *TodoTagRelationsRepository) TransOverwrite(tx *sql.Tx, todoId int, tagIds []int) (err error) {
-	_, err = repo.TransExecute(tx, DeleteTodoTagRelationsState, todoId)
+	_, err = repo.TransExecute(tx, mysql.DeleteTodoTagRelationsState, todoId)
 	if err != nil {
 		fmt.Println(err)
 		log.Println(err)
 		return err
 	}
 	for _, v := range tagIds {
-		_, err = repo.TransExecute(tx, CreateTodoTagRelationsState, todoId, v, time.Now())
+		_, err = repo.TransExecute(tx, mysql.CreateTodoTagRelationsState, todoId, v, time.Now())
 		if err != nil {
 			fmt.Println(err)
 			log.Println(err)
